Remove candidates.rs and refactor RefineCandidate in fitter.rs
- Delete src/grid/candidates.rs file
- Move update_predictions method into TreeGridFitter struct
- Simplify RefineCandidate struct by removing unnecessary fields
- Update find_refine_candidate to use cells instead of curr_leaf_points_idx
- Inline update_predictions function within the TreeGridFitter implementation
